+++ api/holdings/pnl/simple-history.ts
@@ -1,149 +1,1 @@
-import type { VercelRequest, VercelResponse } from '@vercel/node'
-import type {
-  HoldingsEventFetchType,
-  HoldingsEventPaginationMode,
-  HoldingsHistoryTimeframe,
-  VaultVersion
-} from '../../lib/holdings'
-import { checkRateLimit, ensureSchemaInitialized } from '../../lib/holdings'
-
-function simpleHash(str: string): string {
-  const hash = Array.from(str).reduce((currentHash, char) => {
-    const nextHash = (currentHash << 5) - currentHash + char.charCodeAt(0)
-    return nextHash & nextHash
-  }, 0)
-  return Math.abs(hash).toString(36)
-}
-
-function getClientIdentifier(req: VercelRequest): string {
-  const forwarded = req.headers['x-forwarded-for']
-  if (forwarded) {
-    return (Array.isArray(forwarded) ? forwarded[0] : forwarded).split(',')[0].trim()
-  }
-
-  const ua = req.headers['user-agent'] || ''
-  const lang = req.headers['accept-language'] || ''
-  const encoding = req.headers['accept-encoding'] || ''
-  return `fp-${simpleHash(ua + lang + encoding)}`
-}
-
-function isValidAddress(address: string): boolean {
-  return /^0x[a-fA-F0-9]{40}$/.test(address)
-}
-
-function parseHoldingsEventFetchType(value: string | string[] | undefined): HoldingsEventFetchType {
-  return value === 'parallel' ? 'parallel' : 'seq'
-}
-
-function parseHoldingsEventPaginationMode(value: string | string[] | undefined): HoldingsEventPaginationMode {
-  return value === 'all' ? 'all' : 'paged'
-}
-
-function parseHoldingsHistoryTimeframe(value: string | string[] | undefined): HoldingsHistoryTimeframe {
-  return value === 'all' ? 'all' : '1y'
-}
-
-export default async function handler(req: VercelRequest, res: VercelResponse) {
-  res.setHeader('Access-Control-Allow-Origin', '*')
-  res.setHeader('Access-Control-Allow-Methods', 'GET, OPTIONS')
-  res.setHeader('Access-Control-Allow-Headers', 'Content-Type')
-
-  if (req.method === 'OPTIONS') {
-    return res.status(204).end()
-  }
-
-  if (req.method !== 'GET') {
-    return res.status(405).json({ error: 'Method not allowed' })
-  }
-
-  try {
-    await ensureSchemaInitialized()
-  } catch (error) {
-    console.error('Holdings simple PnL history schema initialization error:', error)
-    return res.status(500).json({ error: 'Failed to initialize holdings storage' })
-  }
-
-  const clientId = getClientIdentifier(req)
-  const rateCheck = await checkRateLimit(clientId)
-  if (!rateCheck.allowed) {
-    res.setHeader('Retry-After', String(rateCheck.retryAfter))
-    return res.status(429).json({ error: 'Too many requests', retryAfter: rateCheck.retryAfter })
-  }
-
-  const envioUrl = process.env.ENVIO_GRAPHQL_URL
-  if (!envioUrl) {
-    return res.status(503).json({
-      error: 'Holdings simple PnL history API not configured',
-      details: 'ENVIO_GRAPHQL_URL environment variable is not set. This feature requires a running Envio indexer.'
-    })
-  }
-
-  const {
-    address,
-    chainId: chainIdParam,
-    vault: vaultParam,
-    version: versionParam,
-    fetchType: fetchTypeParam,
-    paginationMode: paginationModeParam,
-    timeframe: timeframeParam
-  } = req.query
-
-  if (!address || typeof address !== 'string') {
-    return res.status(400).json({ error: 'Missing required parameter: address' })
-  }
-
-  if (!isValidAddress(address)) {
-    return res.status(400).json({ error: 'Invalid Ethereum address' })
-  }
-
-  if (vaultParam !== undefined && typeof vaultParam !== 'string') {
-    return res.status(400).json({ error: 'Invalid vault address' })
-  }
-
-  if (vaultParam && !isValidAddress(vaultParam)) {
-    return res.status(400).json({ error: 'Invalid vault address' })
-  }
-
-  if (vaultParam && (!chainIdParam || typeof chainIdParam !== 'string' || !Number.isInteger(Number(chainIdParam)))) {
-    return res.status(400).json({ error: 'Missing or invalid chainId for vault filter' })
-  }
-
-  const version: VaultVersion = versionParam === 'v2' || versionParam === 'v3' ? versionParam : 'all'
-  const fetchType = parseHoldingsEventFetchType(fetchTypeParam)
-  const paginationMode = parseHoldingsEventPaginationMode(paginationModeParam)
-  const timeframe = parseHoldingsHistoryTimeframe(timeframeParam)
-
-  try {
-    const { getHoldingsPnLSimpleHistory } = await import('../../lib/holdings')
-    const history = await getHoldingsPnLSimpleHistory(
-      address,
-      version,
-      fetchType,
-      paginationMode,
-      timeframe,
-      vaultParam,
-      vaultParam ? Number(chainIdParam) : undefined
-    )
-
-    if (history.summary.totalVaults === 0) {
-      return res.status(404).json({ error: 'No holdings found for address' })
-    }
-
-    res.setHeader('Cache-Control', 'public, s-maxage=300, stale-while-revalidate=600')
-    return res.status(200).json(history)
-  } catch (error) {
-    console.error('Holdings simple PnL history error:', error)
-
-    if (process.env.NODE_ENV === 'development') {
-      const message = error instanceof Error ? error.message : String(error)
-      const stack = error instanceof Error ? error.stack : undefined
-      return res.status(502).json({
-        error: 'Failed to fetch holdings simple PnL history',
-        message,
-        stack
-      })
-    }
-
-    return res.status(502).json({ error: 'Failed to fetch holdings simple PnL history' })
-  }
-}
+export { default } from '../protocol-return/history'
