+++ api/lib/holdings/config.ts
@@ -2,7 +2,6 @@ export interface HoldingsConfig {
   readonly envioGraphqlUrl: string
   readonly envioPassword: string
   readonly databaseUrl: string | null
-  readonly ethereumRpcUrl: string
   readonly kongBaseUrl: string
   readonly defillamaBaseUrl: string
   readonly defillamaProBaseUrl: string
@@ -20,9 +19,6 @@ export const config: HoldingsConfig = {
   get databaseUrl() {
     return process.env.DATABASE_URL_PREVIEW ?? process.env.DATABASE_URL ?? null
   },
-  get ethereumRpcUrl() {
-    return process.env.ETHEREUM_RPC_URL ?? 'https://ethereum-rpc.publicnode.com'
-  },
   kongBaseUrl: 'https://kong.yearn.fi',
   defillamaBaseUrl: 'https://coins.llama.fi',
   defillamaProBaseUrl: 'https://pro-api.llama.fi',
@@ -38,8 +34,5 @@ export function validateConfig(): void {
   }
   if (!process.env.DATABASE_URL_PREVIEW && !process.env.DATABASE_URL) {
     console.warn('[Holdings] DATABASE_URL_PREVIEW / DATABASE_URL not set, caching disabled')
-  }
-  if (!process.env.ETHEREUM_RPC_URL) {
-    console.warn('[Holdings] ETHEREUM_RPC_URL not set, using default public mainnet RPC')
   }
 }
