+++ api/lib/holdings/services/debug.ts
@@ -3,7 +3,7 @@ import { AsyncLocalStorage } from 'node:async_hooks'
 export interface HoldingsDebugContext {
   enabled: boolean
   requestId: string
-  route: 'history' | 'breakdown' | 'pnl'
+  route: 'history' | 'breakdown' | 'protocol-return-history'
   address: string
   startedAt: number
   lotsEnabled: boolean
@@ -31,7 +31,7 @@ export function isHoldingsDebugRequested(debugValue?: string | null): boolean {
 }
 
 export function createHoldingsDebugContext(
-  route: 'history' | 'breakdown' | 'pnl',
+  route: 'history' | 'breakdown' | 'protocol-return-history',
   address: string,
   enabled: boolean,
   options?: {
