+++ api/server.ts
@@ -11,17 +11,13 @@ import {
   getHistoricalHoldingsChart,
   getHoldingsActivity,
   getHoldingsBreakdown,
-  getHoldingsPnL,
-  getHoldingsPnLDrilldown,
-  getHoldingsPnLSimple,
-  getHoldingsPnLSimpleHistory,
+  getHoldingsProtocolReturnHistory,
   type HoldingsEventFetchType,
   type HoldingsEventPaginationMode,
   type HoldingsHistoryDenomination,
   type HoldingsHistoryTimeframe,
   initializeSchema,
   isDatabaseEnabled,
-  type UnknownTransferInPnlMode,
   type VaultVersion,
   validateConfig
 } from './lib/holdings'
@@ -162,10 +158,6 @@ function validateInvalidateBody(body: unknown): body is InvalidateRequestBody {
   }
 
   return true
-}
-
-function parseUnknownTransferInPnlMode(value: string | null): UnknownTransferInPnlMode {
-  return value === 'strict' || value === 'zero_basis' || value === 'windfall' ? value : 'windfall'
 }
 
 function parseHoldingsEventFetchType(value: string | null): HoldingsEventFetchType {
@@ -725,169 +717,7 @@ async function handleHoldingsBreakdown(req: Request): Promise<Response> {
   }
 }
 
-async function handleHoldingsPnL(req: Request): Promise<Response> {
-  const url = new URL(req.url)
-  const address = url.searchParams.get('address')
-  const versionParam = url.searchParams.get('version')
-  const debugEnabled =
-    isHoldingsDebugRequested(url.searchParams.get('debug')) || isHoldingsDebugRequested(process.env.HOLDINGS_DEBUG)
-  const debugLotsEnabled = isHoldingsDebugRequested(url.searchParams.get('debugLots'))
-  const debugVault = url.searchParams.get('debugVault')
-  const debugTx = url.searchParams.get('debugTx')
-  const unknownTransferInPnlMode = parseUnknownTransferInPnlMode(url.searchParams.get('unknownMode'))
-  const fetchType = parseHoldingsEventFetchType(url.searchParams.get('fetchType'))
-  const paginationMode = parseHoldingsEventPaginationMode(url.searchParams.get('paginationMode'))
-
-  if (!address) {
-    return Response.json({ error: 'Missing required parameter: address', status: 400 }, { status: 400 })
-  }
-
-  if (!isValidAddress(address)) {
-    return Response.json({ error: 'Invalid Ethereum address', status: 400 }, { status: 400 })
-  }
-
-  const version: VaultVersion = versionParam === 'v2' || versionParam === 'v3' ? versionParam : 'all'
-
-  try {
-    const pnl = await withHoldingsDebugContext(
-      createHoldingsDebugContext('pnl', address, debugEnabled, {
-        lotsEnabled: debugLotsEnabled,
-        vaultFilter: debugVault,
-        txFilter: debugTx
-      }),
-      async () => {
-        debugLog('route', 'started holdings pnl request', {
-          version,
-          unknownTransferInPnlMode,
-          fetchType,
-          paginationMode,
-          debugLotsEnabled,
-          debugVault: debugVault?.toLowerCase() ?? null,
-          debugTx: debugTx?.toLowerCase() ?? null
-        })
-
-        try {
-          const response = await getHoldingsPnL(address, version, unknownTransferInPnlMode, fetchType, paginationMode)
-          debugLog('route', 'completed holdings pnl request', {
-            version,
-            unknownTransferInPnlMode,
-            fetchType,
-            paginationMode,
-            totalVaults: response.summary.totalVaults,
-            totalCurrentValueUsd: response.summary.totalCurrentValueUsd,
-            totalPnlUsd: response.summary.totalPnlUsd,
-            totalEconomicGainUsd: response.summary.totalEconomicGainUsd
-          })
-          return response
-        } catch (error) {
-          debugError('route', 'holdings pnl request failed', error, {
-            version,
-            unknownTransferInPnlMode,
-            fetchType,
-            paginationMode
-          })
-          throw error
-        }
-      }
-    )
-
-    if (pnl.summary.totalVaults === 0) {
-      return Response.json({ error: 'No holdings found for address', status: 404 }, { status: 404 })
-    }
-
-    return Response.json(pnl, {
-      headers: {
-        'Cache-Control': 'public, s-maxage=300, stale-while-revalidate=600'
-      }
-    })
-  } catch (error) {
-    console.error('Error fetching holdings PnL:', error)
-    const message = error instanceof Error ? error.message : String(error)
-    const stack = error instanceof Error ? error.stack : undefined
-    return Response.json({ error: 'Failed to fetch holdings PnL', message, stack, status: 502 }, { status: 502 })
-  }
-}
-
-async function handleHoldingsPnLSimple(req: Request): Promise<Response> {
-  const url = new URL(req.url)
-  const address = url.searchParams.get('address')
-  const versionParam = url.searchParams.get('version')
-  const debugEnabled =
-    isHoldingsDebugRequested(url.searchParams.get('debug')) || isHoldingsDebugRequested(process.env.HOLDINGS_DEBUG)
-  const debugLotsEnabled = isHoldingsDebugRequested(url.searchParams.get('debugLots'))
-  const debugVault = url.searchParams.get('debugVault')
-  const debugTx = url.searchParams.get('debugTx')
-  const fetchType = parseHoldingsEventFetchType(url.searchParams.get('fetchType'))
-  const paginationMode = parseHoldingsEventPaginationMode(url.searchParams.get('paginationMode'))
-
-  if (!address) {
-    return Response.json({ error: 'Missing required parameter: address', status: 400 }, { status: 400 })
-  }
-
-  if (!isValidAddress(address)) {
-    return Response.json({ error: 'Invalid Ethereum address', status: 400 }, { status: 400 })
-  }
-
-  const version: VaultVersion = versionParam === 'v2' || versionParam === 'v3' ? versionParam : 'all'
-
-  try {
-    const pnl = await withHoldingsDebugContext(
-      createHoldingsDebugContext('pnl-simple', address, debugEnabled, {
-        lotsEnabled: debugLotsEnabled,
-        vaultFilter: debugVault,
-        txFilter: debugTx
-      }),
-      async () => {
-        debugLog('route', 'started holdings simple pnl request', {
-          version,
-          fetchType,
-          paginationMode,
-          debugLotsEnabled,
-          debugVault: debugVault?.toLowerCase() ?? null,
-          debugTx: debugTx?.toLowerCase() ?? null
-        })
-
-        try {
-          const response = await getHoldingsPnLSimple(address, version, fetchType, paginationMode)
-          debugLog('route', 'completed holdings simple pnl request', {
-            version,
-            fetchType,
-            paginationMode,
-            totalVaults: response.summary.totalVaults,
-            baselineWeightUsd: response.summary.baselineWeightUsd,
-            growthWeightUsd: response.summary.growthWeightUsd,
-            protocolReturnPct: response.summary.protocolReturnPct
-          })
-          return response
-        } catch (error) {
-          debugError('route', 'holdings simple pnl request failed', error, {
-            version,
-            fetchType,
-            paginationMode
-          })
-          throw error
-        }
-      }
-    )
-
-    if (pnl.summary.totalVaults === 0) {
-      return Response.json({ error: 'No holdings found for address', status: 404 }, { status: 404 })
-    }
-
-    return Response.json(pnl, {
-      headers: {
-        'Cache-Control': 'public, s-maxage=300, stale-while-revalidate=600'
-      }
-    })
-  } catch (error) {
-    console.error('Error fetching holdings simple PnL:', error)
-    const message = error instanceof Error ? error.message : String(error)
-    const stack = error instanceof Error ? error.stack : undefined
-    return Response.json({ error: 'Failed to fetch holdings simple PnL', message, stack, status: 502 }, { status: 502 })
-  }
-}
-
-async function handleHoldingsPnLSimpleHistory(req: Request): Promise<Response> {
+async function handleHoldingsProtocolReturnHistory(req: Request): Promise<Response> {
   const url = new URL(req.url)
   const address = url.searchParams.get('address')
   const chainIdParam = url.searchParams.get('chainId')
@@ -922,13 +752,13 @@ async function handleHoldingsPnLSimpleHistory(req: Request): Promise<Response> {
 
   try {
     const history = await withHoldingsDebugContext(
-      createHoldingsDebugContext('pnl-simple-history', address, debugEnabled, {
+      createHoldingsDebugContext('protocol-return-history', address, debugEnabled, {
         lotsEnabled: debugLotsEnabled,
         vaultFilter: debugVault,
         txFilter: debugTx
       }),
       async () => {
-        debugLog('route', 'started holdings simple pnl history request', {
+        debugLog('route', 'started holdings protocol return history request', {
           version,
           timeframe,
           vault: vault?.toLowerCase() ?? null,
@@ -941,7 +771,7 @@ async function handleHoldingsPnLSimpleHistory(req: Request): Promise<Response> {
         })
 
         try {
-          const response = await getHoldingsPnLSimpleHistory(
+          const response = await getHoldingsProtocolReturnHistory(
             address,
             version,
             fetchType,
@@ -950,7 +780,7 @@ async function handleHoldingsPnLSimpleHistory(req: Request): Promise<Response> {
             vault ?? undefined,
             vault !== null ? Number(chainIdParam) : undefined
           )
-          debugLog('route', 'completed holdings simple pnl history request', {
+          debugLog('route', 'completed holdings protocol return history request', {
             version,
             timeframe,
             vault: vault?.toLowerCase() ?? null,
@@ -962,7 +792,7 @@ async function handleHoldingsPnLSimpleHistory(req: Request): Promise<Response> {
           })
           return response
         } catch (error) {
-          debugError('route', 'holdings simple pnl history request failed', error, {
+          debugError('route', 'holdings protocol return history request failed', error, {
             version,
             timeframe,
             vault: vault?.toLowerCase() ?? null,
@@ -985,118 +815,11 @@ async function handleHoldingsPnLSimpleHistory(req: Request): Promise<Response> {
       }
     })
   } catch (error) {
-    console.error('Error fetching holdings simple PnL history:', error)
+    console.error('Error fetching holdings protocol return history:', error)
     const message = error instanceof Error ? error.message : String(error)
     const stack = error instanceof Error ? error.stack : undefined
     return Response.json(
-      { error: 'Failed to fetch holdings simple PnL history', message, stack, status: 502 },
-      { status: 502 }
-    )
-  }
-}
-
-async function handleHoldingsPnLDrilldown(req: Request): Promise<Response> {
-  const url = new URL(req.url)
-  const address = url.searchParams.get('address')
-  const vault = url.searchParams.get('vault')
-  const versionParam = url.searchParams.get('version')
-  const debugEnabled =
-    isHoldingsDebugRequested(url.searchParams.get('debug')) || isHoldingsDebugRequested(process.env.HOLDINGS_DEBUG)
-  const debugLotsEnabled = isHoldingsDebugRequested(url.searchParams.get('debugLots'))
-  const debugVault = url.searchParams.get('debugVault') ?? vault
-  const debugTx = url.searchParams.get('debugTx')
-  const unknownTransferInPnlMode = parseUnknownTransferInPnlMode(url.searchParams.get('unknownMode'))
-  const fetchType = parseHoldingsEventFetchType(url.searchParams.get('fetchType'))
-  const paginationMode = parseHoldingsEventPaginationMode(url.searchParams.get('paginationMode'))
-
-  if (!address) {
-    return Response.json({ error: 'Missing required parameter: address', status: 400 }, { status: 400 })
-  }
-
-  if (!isValidAddress(address)) {
-    return Response.json({ error: 'Invalid Ethereum address', status: 400 }, { status: 400 })
-  }
-
-  if (vault !== null && !isValidAddress(vault)) {
-    return Response.json({ error: 'Invalid vault address', status: 400 }, { status: 400 })
-  }
-
-  const version: VaultVersion = versionParam === 'v2' || versionParam === 'v3' ? versionParam : 'all'
-
-  try {
-    const pnl = await withHoldingsDebugContext(
-      createHoldingsDebugContext('pnl', address, debugEnabled, {
-        lotsEnabled: debugLotsEnabled,
-        vaultFilter: debugVault,
-        txFilter: debugTx
-      }),
-      async () => {
-        debugLog('route', 'started holdings pnl drilldown request', {
-          version,
-          unknownTransferInPnlMode,
-          fetchType,
-          paginationMode,
-          vault: vault?.toLowerCase() ?? null,
-          debugLotsEnabled,
-          debugVault: debugVault?.toLowerCase() ?? null,
-          debugTx: debugTx?.toLowerCase() ?? null
-        })
-
-        try {
-          const response = await getHoldingsPnLDrilldown(
-            address,
-            version,
-            unknownTransferInPnlMode,
-            fetchType,
-            paginationMode,
-            vault
-          )
-          debugLog('route', 'completed holdings pnl drilldown request', {
-            version,
-            unknownTransferInPnlMode,
-            fetchType,
-            paginationMode,
-            vault: vault?.toLowerCase() ?? null,
-            totalVaults: response.summary.totalVaults,
-            totalCurrentValueUsd: response.summary.totalCurrentValueUsd,
-            totalPnlUsd: response.summary.totalPnlUsd,
-            totalEconomicGainUsd: response.summary.totalEconomicGainUsd
-          })
-          return response
-        } catch (error) {
-          debugError('route', 'holdings pnl drilldown request failed', error, {
-            version,
-            unknownTransferInPnlMode,
-            fetchType,
-            paginationMode,
-            vault: vault?.toLowerCase() ?? null
-          })
-          throw error
-        }
-      }
-    )
-
-    if (pnl.summary.totalVaults === 0) {
-      return Response.json(
-        {
-          error: vault ? 'No matching holdings found for address and vault' : 'No holdings found for address',
-          status: 404
-        },
-        { status: 404 }
-      )
-    }
-
-    return Response.json(pnl, {
-      headers: {
-        'Cache-Control': 'public, s-maxage=300, stale-while-revalidate=600'
-      }
-    })
-  } catch (error) {
-    console.error('Error fetching holdings PnL drilldown:', error)
-    const message = error instanceof Error ? error.message : String(error)
-    const stack = error instanceof Error ? error.stack : undefined
-    return Response.json(
-      { error: 'Failed to fetch holdings PnL drilldown', message, stack, status: 502 },
+      { error: 'Failed to fetch holdings protocol return history', message, stack, status: 502 },
       { status: 502 }
     )
   }
@@ -1243,20 +966,11 @@ async function main() {
           return withCors(await handleHoldingsBreakdown(req))
         }
 
-        if (url.pathname === '/api/holdings/pnl/drilldown') {
-          return withCors(await handleHoldingsPnLDrilldown(req))
-        }
-
-        if (url.pathname === '/api/holdings/pnl/simple') {
-          return withCors(await handleHoldingsPnLSimple(req))
-        }
-
-        if (url.pathname === '/api/holdings/pnl/simple-history') {
-          return withCors(await handleHoldingsPnLSimpleHistory(req))
-        }
-
-        if (url.pathname === '/api/holdings/pnl') {
-          return withCors(await handleHoldingsPnL(req))
+        if (
+          url.pathname === '/api/holdings/protocol-return/history' ||
+          url.pathname === '/api/holdings/pnl/simple-history'
+        ) {
+          return withCors(await handleHoldingsProtocolReturnHistory(req))
         }
 
         if (url.pathname === '/api/holdings/chores') {
@@ -1318,18 +1032,17 @@ async function main() {
         )
       }
     },
-    port: 3001,
+    port: API_SERVER_PORT,
     idleTimeout: 120
   })
 
-  console.log('🚀 API server running on http://localhost:3001')
-  console.log('📊 Holdings API: http://localhost:3001/api/holdings/history?address=0x...')
-  console.log('🗂️ Holdings Activity API: http://localhost:3001/api/holdings/activity?address=0x...')
-  console.log('🧩 Holdings Breakdown API: http://localhost:3001/api/holdings/breakdown?address=0x...')
-  console.log('💹 PnL API: http://localhost:3001/api/holdings/pnl?address=0x...')
-  console.log('📈 Simple PnL API: http://localhost:3001/api/holdings/pnl/simple?address=0x...')
-  console.log('📊 Simple PnL History API: http://localhost:3001/api/holdings/pnl/simple-history?address=0x...')
-  console.log('🧾 PnL Drilldown API: http://localhost:3001/api/holdings/pnl/drilldown?address=0x...')
+  console.log(`🚀 API server running on http://localhost:${API_SERVER_PORT}`)
+  console.log(`📊 Holdings API: http://localhost:${API_SERVER_PORT}/api/holdings/history?address=0x...`)
+  console.log(`🗂️ Holdings Activity API: http://localhost:${API_SERVER_PORT}/api/holdings/activity?address=0x...`)
+  console.log(`🧩 Holdings Breakdown API: http://localhost:${API_SERVER_PORT}/api/holdings/breakdown?address=0x...`)
+  console.log(
+    `📈 Protocol Return History API: http://localhost:${API_SERVER_PORT}/api/holdings/protocol-return/history?address=0x...`
+  )
 }
 
 main().catch((error) => {
